refactor(ui): extract settings/data/dom.ts — pure DOM helpers for tooltip + focus
